requirements.txt
gitpython
pygithub
imap-tools
anthropic
